#535: Move heading from health check view to DivisionHealth component
