+++ CourageScores/ClientApp/src/components/division_health/ViewHealthCheck.test.js
@@ -29,77 +29,6 @@ describe('DivisionHealth', () => {
             (<ViewHealthCheck result={result} />));
     }
 
-    function assertHeading(text, className) {
-        const heading = context.container.querySelector('h3');
-        expect(heading.textContent).toEqual(text);
-        expect(heading.className).toContain(className);
-    }
-
-    it('when success and no errors or warnings should show healthy', async () => {
-        await renderComponent({
-            success: true,
-            errors: [],
-            warnings: [],
-            messages: [],
-            checks: {},
-        });
-
-        expect(reportedError).toBeNull();
-        assertHeading('Status: Healthy', 'text-success');
-    });
-
-    it('when success and some errors should show unhealthy', async () => {
-        await renderComponent({
-            success: true,
-            errors: [ 'some error' ],
-            warnings: [],
-            messages: [],
-            checks: {},
-        });
-
-        expect(reportedError).toBeNull();
-        assertHeading('Status: Unhealthy', 'text-warning');
-    });
-
-    it('when success and some warnings should show unhealthy', async () => {
-        await renderComponent({
-            success: true,
-            errors: [],
-            warnings: [ 'warning' ],
-            messages: [],
-            checks: {},
-        });
-
-        expect(reportedError).toBeNull();
-        assertHeading('Status: Unhealthy', 'text-warning');
-    });
-
-    it('when success and some messages should show healthy', async () => {
-        await renderComponent({
-            success: true,
-            errors: [],
-            warnings: [],
-            messages: [ 'message' ],
-            checks: {},
-        });
-
-        expect(reportedError).toBeNull();
-        assertHeading('Status: Healthy', 'text-success');
-    });
-
-    it('when unsuccess and no errors, warnings or messages should show unhealthy', async () => {
-        await renderComponent({
-            success: false,
-            errors: [],
-            warnings: [],
-            messages: [],
-            checks: {},
-        });
-
-        expect(reportedError).toBeNull();
-        assertHeading('Status: Unhealthy', 'text-warning');
-    });
-
     it('should render successful check', async () => {
         await renderComponent({
             success: false,
